pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["piotrstempkowski <pstempkowski19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
ipython = "^8.12.0"
django-cors-headers = "^3.14.0"
psycopg2 = "^2.9.5"
psycopg2-binary = "^2.9.5"
factory-boy = "^3.2.1"
django-filter = "^23.1"
drf-spectacular = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
